+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 4.5.1
 ----------------
 - Move Header head action hook outisde page condition in Warehouse.php
 - Fix SQL error, do not allow N/A in Timeframe date inputs in Calendar.php
+- JS Add ajaxPopState() function & Fix logout when back button & URL is Modules.php in warehouse.js
 
 Changes in 4.5
 --------------