+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 4.5.1
 - JS Add ajaxPopState() function & Fix logout when back button & URL is Modules.php in warehouse.js
 - Upload photo when no other fields are posted in Student.php & User.php
 - Make makeProfile() function reusable & better display in GetStaffList.fnc.php
+- Fix SaveTemplate dynamic modname in Template.fnc.php
 
 Changes in 4.5
 --------------